--- configs/bootstrap/dex/web/templates/error.html
@@ -1,8 +1,0 @@
-{{ template "header.html" . }}
-
-<div class="theme-panel">
-  <h2 class="theme-heading">{{ .ErrType }}</h2>
-  <p>{{ .ErrMsg }}</p>
-</div>
-
-{{ template "footer.html" . }}
