--- configs/bootstrap/dex/web/templates/footer.html
@@ -1,3 +1,0 @@
-    </div>
-  </body>
-</html>
