--- configs/bootstrap/dex/web/templates/login.html
@@ -1,19 +1,0 @@
-{{ template "header.html" . }}
-
-<div class="theme-panel">
-  <h2 class="theme-heading">Log in to {{ issuer }} </h2>
-  <div>
-    {{ range $c := .Connectors }}
-      <div class="theme-form-row">
-        <a href="{{ $c.URL }}" target="_self">
-          <button class="dex-btn theme-btn-provider">
-            <span class="dex-btn-icon dex-btn-icon--{{ $c.ID }}"></span>
-            <span class="dex-btn-text">Log in with {{ $c.Name }}</span>
-          </button>
-        </a>
-      </div>
-    {{ end }}
-  </div>
-</div>
-
-{{ template "footer.html" . }}
