--- configs/bootstrap/dex/web/templates/password.html
@@ -1,35 +1,0 @@
-{{ template "header.html" . }}
-
-<div class="theme-panel">
-  <h2 class="theme-heading">Log in to Your Account</h2>
-  <form method="post" action="{{ .PostURL }}">
-    <div class="theme-form-row">
-      <div class="theme-form-label">
-        <label for="userid">{{ .UsernamePrompt }}</label>
-      </div>
-	  <input tabindex="1" required id="login" name="login" type="text" class="theme-form-input" placeholder="{{ .UsernamePrompt | lower }}" {{ if .Username }} value="{{ .Username }}" {{ else }} autofocus {{ end }}/>
-    </div>
-    <div class="theme-form-row">
-      <div class="theme-form-label">
-        <label for="password">Password</label>
-      </div>
-	  <input tabindex="2" required id="password" name="password" type="password" class="theme-form-input" placeholder="password" {{ if .Invalid }} autofocus {{ end }}/>
-    </div>
-
-    {{ if .Invalid }}
-      <div id="login-error" class="dex-error-box">
-        Invalid {{ .UsernamePrompt }} and password.
-      </div>
-    {{ end }}
-
-    <button tabindex="3" id="submit-login" type="submit" class="dex-btn theme-btn--primary">Login</button>
-
-  </form>
-  {{ if .BackLink }}
-  <div class="theme-link-back">
-    <a class="dex-subtle-text" href="javascript:history.back()">Select another login method.</a>
-  </div>
-  {{ end }}
-</div>
-
-{{ template "footer.html" . }}
