--- configs/bootstrap/dex/web/themes/tectonic/styles.css
@@ -1,106 +1,0 @@
-.theme-body {
-  background-color: #1C425B;
-  color: #333;
-  font-family: 'Source Sans Pro', Helvetica, sans-serif;
-}
-
-.theme-navbar {
-  display: none;
-}
-
-.theme-navbar__logo-wrap {
-  display: inline-block;
-  height: 100%;
-  overflow: hidden;
-  padding: 10px 15px;
-  width: 300px;
-}
-
-.theme-navbar__logo {
-  height: 100%;
-  max-height: 25px;
-}
-
-.theme-heading {
-  font-size: 20px;
-  font-weight: 500;
-  padding-bottom: 10px;
-  margin-top: 0;
-}
-
-.theme-panel {
-  background-color: #fff;
-  box-shadow: 0 5px 15px rgba(0, 0, 0, 0.5);
-  padding: 30px;
-}
-
-.theme-btn-provider {
-  background-color: #fff;
-  color: #333;
-  min-width: 250px;
-}
-
-.theme-btn-provider:hover {
-  color: #999;
-}
-
-.theme-btn--primary {
-  background-color: #2fc98e;
-  border: none;
-  color: #fff;
-  min-width: 200px;
-  padding: 6px 12px;
-}
-
-.theme-btn--primary:hover {
-  background-color: #25a071;
-  color: #fff;
-}
-
-.theme-btn--success {
-  background-color: #2FC98E;
-  color: #fff;
-  width: 250px;
-}
-
-.theme-btn--success:hover {
-  background-color: #49E3A8;
-}
-
-.theme-form-row {
-  display: block;
-  margin: 20px auto;
-}
-
-.theme-form-input {
-  display: block;
-  height: 36px;
-  padding: 6px 12px;
-  font-size: 14px;
-  line-height: 1.42857143;
-  color: #666;
-  border: 1px solid #CCC;
-  border-radius: 4px;
-  box-shadow: inset 0 1px 1px rgba(0, 0, 0, 0.075);
-  width: 250px;
-  margin: auto;
-}
-
-.theme-form-input:focus,
-.theme-form-input:active {
-  outline: none;
-  border-color: #66AFE9;
-}
-
-.theme-form-label {
-  width: 250px;
-  margin: 4px auto;
-  text-align: left;
-  position: relative;
-  font-size: 13px;
-  font-weight: 600;
-}
-
-.dex-container {
-  margin: 60px auto;
-}
