--- configs/bootstrap/postgres/init-dex-db.sh
@@ -1,8 +1,0 @@
-#!/bin/bash
-set -e
-
-psql -v ON_ERROR_STOP=1 --username "$POSTGRES_USER" <<-EOSQL
-	CREATE USER dex;
-  CREATE DATABASE dex;
-	GRANT ALL PRIVILEGES ON DATABASE dex TO dex;
-EOSQL
