+++ plugins/utils.go
@@ -6,51 +6,6 @@ import (
 	"github.com/extemporalgenome/slug"
 )
 
-type KeyValue struct {
-	Key   string `json:"key"`
-	Value string `json:"value"`
-}
-
-func ConvertMapStringInterfaceToKV(formSpecMap map[string]interface{}, kv *[]KeyValue) error {
-	for key, value := range formSpecMap {
-		*kv = append(*kv, KeyValue{Key: key, Value: value.(string)})
-	}
-	return nil
-}
-
-func ConvertKVToMapStringInterface(kv []KeyValue, formSpecMap *map[string]interface{}) error {
-	formMap := *formSpecMap
-
-	for _, keyValue := range kv {
-		tmpKv := keyValue
-		formMap[tmpKv.Key] = &tmpKv.Value
-	}
-
-	formSpecMap = &formMap
-	return nil
-}
-
-func ConvertKVToMapStringString(kv []KeyValue, formSpecMap *map[string]*string) error {
-
-	formMap := *formSpecMap
-
-	for _, keyValue := range kv {
-		tmpKv := keyValue
-		formMap[tmpKv.Key] = &tmpKv.Value
-	}
-
-	formSpecMap = &formMap
-
-	return nil
-}
-
-func ConvertMapStringStringToKV(formSpecMap map[string]*string, kv *[]KeyValue) error {
-	for key, value := range formSpecMap {
-		*kv = append(*kv, KeyValue{Key: key, Value: *value})
-	}
-	return nil
-}
-
 func GetRegexParams(regEx, url string) (paramsMap map[string]string) {
 	var compRegEx = regexp.MustCompile(regEx)
 	match := compRegEx.FindStringSubmatch(url)
